change api/user (method=GET) > payload
