+++ analysis/test/configparser/example.ini
@@ -1,0 +1,13 @@
+[DEFAULT]
+serveraliveinterval = 45
+compression = yes
+compressionlevel = 9
+forwardx11 = yes
+
+[bitbucket.org]
+user = hg
+
+[topsecret.server.com]
+port = 50022
+forwardx11 = no
+
