+++ analysis/test/python/my.conf
@@ -1,0 +1,8 @@
+[connector_python]
+host=127.0.0.1
+database=taipei_trip
+user=root
+password=0973
+port=3306
+[mysqld]
+secure_file_priv=''
