+++ app.py
@@ -14,6 +14,9 @@ app.register_blueprint(search_bp)
 
 from api.user.user import user_system_bp
 app.register_blueprint(user_system_bp)
+from api.booking.booking import booking_bp
+app.register_blueprint(booking_bp)
+
 
 # Pages
 @app.route("/")
