--- test/attraction.css
@@ -1,158 +1,0 @@
-* {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    
-    font-family: Noto Sans TC, Arial, Helvetica, sans-serif;
-    font-size: 16px;
-    font-weight: 400;
-}
-
-a {
-    text-decoration: none;
-}
-
-li{
-    list-style: none;
-}
-
-img { 
-    display: block;
-}
-
-/* >=1200px */
-@media screen and (min-width: 1200px) {
-    /* ==================== header start ==================== */
-    .type_area {
-        width: 1200px;
-    }
-
-    header {
-        position: sticky;
-        top: 0;
-
-        border-bottom: 1px solid #E8E8E8;
-        width: 100%;
-        z-index: 1;
-        background-color: #fff;
-    }
-    
-    header nav {
-        display: flex;
-        justify-content: space-between;
-
-        margin: 0 auto;
-        height: 54px;
-        /* background-color: sandybrown; */
-    }
-    
-    nav h1 a {
-        width: 150px;
-        font-size: 30px;
-        font-weight: 700;
-        line-height: 54px;
-        color: #448899;
-        /* background-color: seagreen; */
-    }
-    
-    nav ul {
-        width: 175px;
-        /* background-color: blue; */
-    }
-
-    nav ul a {
-        float: left;
-        padding-right: 20px;
-        line-height: 54px;
-        color: #666;
-    }
-    /* ==================== header end ==================== */
-
-    /* ==================== main start ==================== */
-    .top_box {
-        margin: 0 auto;
-        height: 450px;
-        background-color: #448899;
-    }
-    /* ----------- left_box ----------- */
-    .left_box {
-        float: left;
-        width: 50%;
-        height: 100%;
-        background-color: #994411;
-    }
-    
-    .control_box {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-
-        width: 100%;
-        height: 100%;
-        background-color: #757575;
-    }
-    .control_box span {
-        display: block;
-        width: 36px;
-        height: 36px;
-        background-color: #E8E8E8;
-    }
-
-    .control_box span:first-of-type {
-        background: url(../imgs/btn_leftArrow.png);
-    }
-    .control_box span:last-of-type {
-        background: url(../imgs/btn_rightArrow.png);
-    }
-
-    .control_box ul {
-        align-self: flex-end;
-        display: flex;
-
-        margin-bottom: 13px;
-        /* background-color: #448899; */
-    }
-
-    .control_box ul li {
-        margin-right: 12px;
-        border: 1px solid #fff;
-        border-radius: 12px;
-        width: 12px;
-        height: 12px;
-        background-color: #fff;
-    }
-    .control_box ul :nth-child(2) {
-        background-color: #000;
-    }
-
-    /* ----------- right_box ----------- */
-    .right_box {
-        float: left;
-        width: 50%;
-        height: 100%;
-        background-color: #114499;
-    }
-
-    .bottom_box {
-        margin: 0 auto;
-        height: 450px;
-        background-color: #998844;
-    }
-    /* ==================== main end ==================== */
-    
-    /* ==================== footer start ==================== */
-    footer {
-        display: flex;
-        justify-content: center;
-        width: 100%;
-        height: 104px;
-        background-color: #757575;
-    }
-
-    .copyright {
-        line-height: 104px;
-        color: #fff;
-    }
-    /* ==================== footer end ==================== */
-
-}
